% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coverage.R
\name{check_coverage}
\alias{check_coverage}
\title{Check if a point is in covered area}
\usage{
check_coverage(point, api_token = Sys.getenv("CITYMAPPER_API_TOKEN"))
}
\arguments{
\item{point}{Geographical coordinates of the point in WGS84 \code{'<latitude>,<longitude>'} format.}

\item{api_token}{Citymapper's api key (chr).}
}
\value{
A tibble containing boolean responses for each point.
}
\description{
Checks if a given point falls within Citymapper's covered areas.
It is good practice to refresh this value regularly as covered areas might change.
}
\examples{
\dontrun{
check_coverage("45.448643,9.207645")
}

}
