% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coverage_multi.R
\name{check_coverage_multi}
\alias{check_coverage_multi}
\title{Checks if multiple points are in covered area}
\usage{
check_coverage_multi(points, api_token = Sys.getenv("CITYMAPPER_API_TOKEN"))
}
\arguments{
\item{points}{Dataframe containing geographical coordinates of the start point in WGS84 \code{'<latitude>,<longitude>'} format.
Columns should be set as \code{id (optional)} and \code{coord}.}

\item{api_token}{Citymapper's api key (chr).}
}
\value{
A tibble containing boolean responses for each point. IDs column passed in \code{points} will be mirrored back in response.
}
\description{
Checks if  multiple points fall within Citymapper's covered areas.
It is good practice to refresh this values regularly as covered areas might change.
}
\examples{
\dontrun{
# Make dataframes with coordinates and (optional) ids

}

}
\seealso{
\code{\link{check_coverage}}
}
