% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citymappR.R
\docType{package}
\name{citymappR}
\alias{citymappR}
\title{citymappR: Wrapper for Citymapper's API}
\description{
A package to interact with Citymapper API.

The package lets you compute travel time between two lat/lon points.
  As of \code{0.1.1} the package allows to download \href{https://citymapper.com/cmi}{
  Citymapper's city mobility index}
}
\keyword{internal}
