% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citymappr_setup.R
\name{citymappr_setup}
\alias{citymappr_setup}
\title{Setup env variables for Citymapper's API access}
\usage{
citymappr_setup(config_file = "~/.citymappr", api_token = "", echo = FALSE)
}
\arguments{
\item{config_file}{a configuration file in DCF format. See \link{read.dcf}.}

\item{api_token}{Citymapper's api key (chr).}

\item{echo}{display the \code{api_token} (bool), initally set as \code{FALSE}}
}
\value{
An env variable
}
\description{
Initialize the API key \code{citymappR} will send in every request for authentication.
If a file is found at the locatioin specified in \code{config_file}, the values will be used.
}
\details{
\preformatted{
api_token: YOUR_TOKEN_STRING
}
}
