% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mob_idx.R
\name{get_mob_idx}
\alias{get_mob_idx}
\title{Get Citymapper's mobility index}
\usage{
get_mob_idx(
  start_date = "2020-01-01",
  end_date = Sys.Date(),
  city = NA,
  cache = TRUE,
  weekly = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{start_date}{First observed day with format YYYY-MM-DD}

\item{end_date}{Last observed day with format YYYY-MM-DD}

\item{city}{(optional) A vector containing specific cities to filter for (e.g "milan").
This obviously follow Citymapper's \href{https://citymapper.com/cities?lang=en}{coverage}.}

\item{cache}{(bool) cache results to filesystem}

\item{weekly}{Get weekly data (default is daily)}

\item{verbose}{Show download progress bar}
}
\value{
A data frame containing the following columns:
  \itemize{
    \item \code{big_change}: (bool) Whether there was a consisten change with previous day.
    \item \code{cange_from_previous}: Difference with previous day.
    \item \code{date}: Observation date
    \item \code{type}: Weekly or daily observation.
    \item \code{region_id}: Region slug.
    \item \code{value}: Index value.
    \item \code{city_name}: Name of oberved city
  }
}
\description{
As of mid-March Citymapper has launched a City Mobility index to observe which
cities are following lockdown measures during COVID-19 pandemic. Data is
calculated comparing the number of trips planned on the app to a typical usage
  period. \href{https://citymapper.com/cmi/about}{Here} for a detailed description.
}
