% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_travel_time.R
\name{get_travel_time}
\alias{get_travel_time}
\title{Get travel time between two points}
\usage{
get_travel_time(
  start_coord,
  end_coord,
  time = get_iso8601_ts(),
  time_type = "arrival",
  api_token = Sys.getenv("CITYMAPPER_API_TOKEN")
)
}
\arguments{
\item{start_coord}{Geographical coordinates of the start point in WGS84 '<latitude>,<longitude>' format.}

\item{end_coord}{Geographical coordinates of the arrival point in WGS84 '<latitude>,<longitude>' format.}

\item{time}{A date & time in ISO-8601 format (e.g \code{2014-11-06T19:00:02-0500}).
If omitted travel time will be computed for travel at the time of the request.}

\item{time_type}{Required if `time` is provided. At the moment the only defined type
is `arrival`, it computes the travel time for arriving at `end_coord` at the given time.}

\item{api_token}{Citymapper's api key (chr).}
}
\value{
Estimated travel time in minutes (int).
}
\description{
Computes estimated travel time between two points using public transport.
If \code{time} and \code{time_type} are not specified, time of travel will be assumed
    to be the same time at which the request is made.
}
\examples{
\dontrun{
get_travel_time(start_coord = "45.448643,9.207645",
                end_coord = "45.452349,9.180225")
}

}
\author{
Andrea Dodet, \email{an.dodet@gmail.com}
}
